% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy.R
\name{efficacy}
\alias{efficacy}
\title{Efficacy Stoping bound}
\usage{
efficacy(N, n1, p0, p1, alpha, beta)
}
\arguments{
\item{N}{maximum sample size}

\item{n1}{the sample size to start interim monitoring}

\item{p0}{null response rate}

\item{p1}{target response rate}

\item{alpha}{alpha}

\item{beta}{beta}
}
\description{
This function generats an efficacy stoping bound giving the parameters p and b.
}
\examples{
efficacy(N=44, n1=1, p0=0.6, p1=0.8,alpha=0.05, beta=0.2)
}
